Move updateSeverLock and version to Monitor
