+++ server/rooms.js
@@ -961,7 +961,7 @@ class GlobalRoom extends BasicRoom {
 		if (Config.autolockdown && Rooms.global.lockdown === true && Rooms.global.battleCount === 0) {
 			// The server is in lockdown, the final battle has finished, and the option is set
 			// so we will now automatically kill the server here if it is not updating.
-			if (Chat.updateServerLock) {
+			if (Monitor.updateServerLock) {
 				this.notifyRooms(notifyPlaces, `|html|<div class="broadcast-red"><b>Automatic server lockdown kill canceled.</b><br /><br />The server tried to automatically kill itself upon the final battle finishing, but the server was updating while trying to kill itself.</div>`);
 				return;
 			}
